Consolidated the two services
