--- invoicing/1.2.3/services/cronJobs/index.js
@@ -1,42 +1,0 @@
-/*
-   Copyright 2016 Yupana systems, Inc
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-       http://www.apache.org/licenses/LICENSE-2.0
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-'use strict';
-
-/* eslint-disable no-console */
-
-exports.$oninitService = {
-    name: 'oninitService',
-    description: 'start a cron job to print console logs',
-    scope: 'public',
-    restricted: false,
-    args: [
-        {
-            name: "serv",
-            type: "YServer",
-            mandatory: false,
-            description: "used to register task"
-        }
-    ],
-    output: {
-        data: "?"
-    },
-    note: "",
-    example: ''
-};
-exports.oninitService = function (serv) {
-    console.log('cron job started ');
-    serv.service.registerTask('cronJob1', 10, () => {
-        return Promise.resolve(console.log('cronJob1 resolved'));
-    });
-    return Promise.resolve();
-};
