--- invoicing/1.2.3/services/cronJobs/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "cronJobs",
-    "version": "1.0.0",
-    "description": "test service",
-    "main": "index.js",
-    "config": {
-      "services": {
-        "resourceConsumption": "singleton",
-        "state": "stateless",
-        "queue": true,
-        "role": {"name": "collaborateur", "version": "*"}
-      }
-    },
-    "homepage": "www.yupana.com",
-    "licence": "commercial",
-    "files": [
-      "index.js"
-    ]
-}
